% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_geom_series.R
\name{sum_geom_series}
\alias{sum_geom_series}
\title{Find the sum of an infinite geometric series.}
\usage{
sum_geom_series(a, r)
}
\arguments{
\item{a}{the starting value of the series}

\item{r}{the ratio between subsequent terms of the series.}
}
\description{
This function finds the sum of entries that make
up a geometric series.
}
